moved functions back to lib/api/links
